+++ src/host/configuration.h
@@ -13,15 +13,27 @@ namespace host
 {
   enum class EnclaveType
   {
-    SGX_RELEASE,
-    SGX_DEBUG,
-    VIRTUAL,
+    RELEASE,
+    DEBUG,
+    VIRTUAL // Deprecated (use EnclavePlatform instead)
   };
   DECLARE_JSON_ENUM(
     EnclaveType,
-    {{EnclaveType::SGX_RELEASE, "Release"},
-     {EnclaveType::SGX_DEBUG, "Debug"},
+    {{EnclaveType::RELEASE, "Release"},
+     {EnclaveType::DEBUG, "Debug"},
      {EnclaveType::VIRTUAL, "Virtual"}});
+
+  enum class EnclavePlatform
+  {
+    SGX,
+    SNP,
+    VIRTUAL,
+  };
+  DECLARE_JSON_ENUM(
+    EnclavePlatform,
+    {{EnclavePlatform::SGX, "SGX"},
+     {EnclavePlatform::SNP, "SNP"},
+     {EnclavePlatform::VIRTUAL, "Virtual"}});
 
   enum class LogFormat
   {
@@ -51,6 +63,7 @@ namespace host
     {
       std::string file;
       EnclaveType type;
+      EnclavePlatform platform;
     };
     Enclave enclave = {};
 
@@ -150,8 +163,9 @@ namespace host
     Command command = {};
   };
 
-  DECLARE_JSON_TYPE(CCHostConfig::Enclave);
+  DECLARE_JSON_TYPE_WITH_OPTIONAL_FIELDS(CCHostConfig::Enclave);
   DECLARE_JSON_REQUIRED_FIELDS(CCHostConfig::Enclave, type, file);
+  DECLARE_JSON_OPTIONAL_FIELDS(CCHostConfig::Enclave, platform);
 
   DECLARE_JSON_TYPE_WITH_OPTIONAL_FIELDS(CCHostConfig::OutputFiles);
   DECLARE_JSON_REQUIRED_FIELDS(CCHostConfig::OutputFiles);