+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Upgraded OpenEnclave to 0.18.4.
 
+### Changed
+
+- `enclave.type` configuration entry now only supports `Debug` or `Release`. Trusted Execution Environment platform should be specified via new `enclave.platform` configuration entry (`SGX`, `SNP` or `Virtual`) (#4569).
+
 ### Fixed
 
 - Fix issue with large snapshots that may cause node crash on startup (join/recover) if configured stack size was too low (#4566).